simplify styles, also get rid of helvetica
helvetica on ubuntu/linux was mapped to some weird font which has
a few extra pixels of padding under each character.  It made
trying to align text inside a box virtually impossible.  Chrome
on linux ignores helvetica declaration (instead of mapping it to
some stupid font) and just uses arial instead.  Arial looks fine
me so just use that.  Now the page numbers in the pagers are in
the middle on both firefox and chrome