try to fix redirects
When goat counter links to pages it strips trailing slashes.  If I think click
a link in goat counter to a page that had the slash stripped, nginx generates
a redirect to the page with the slash.  However, this redirect includes the
port number because the redirect is actually happening in the massassi-static
container/config, not the main one.  This results in an invalid URL (with
the port number) because these port numbers are internal-only and not exposed
to the internet.  This change makes it so that redirect doesn't inclue the
port number anymore.